[FIX]Fixed general formatting, updated manifest files, updated some model name changes, moved user and purchase_order_line to their own files.
