+++ mindesa_purchase/__manifest__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 {
     "name": "Mindesa: Account and Purchase-related customizations",
     "summary": """Implements custom permissions for site management.""",
@@ -10,14 +9,13 @@
     "author": "Odoo Inc",
     "website": "http://odoo.com",
     "category": "Custom Development",
-    "version": "0.1",
+    "version": "1.0",
     "depends": ["purchase_mrp"],
-    'license': 'OEEL-1',
+    'license': 'OPL-1',
     "data": [
         "data/record_rules.xml",
         "data/actions.xml",
         "views/product_views.xml",
         "views/partner_view_inherit.xml"
-        
     ]
 }
