+++ mindesa_purchase/models/__init__.py
@@ -1,8 +1,8 @@
-# -*- coding: utf-8 -*-
 from . import product
 from . import account
 from . import account_payment
 from . import stock_rule
 from . import purchase_order
-from . import procurement_group
+from . import purchase_order_line
 from . import partner
+from . import user
