+++ mindesa_purchase/models/account.py
@@ -1,11 +1,8 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from odoo import api, fields, models, _
 
 
-class AccountInvoice(models.Model):
-    _inherit = "account.invoice"
+class AccountInvoice(models.TransientModel):
+    _inherit = "account.move"
 
     @api.model
     def create(self, vals):
