+++ mindesa_purchase/models/product.py
@@ -1,6 +1,3 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from odoo import api, fields, models, _
 
 
@@ -12,10 +9,4 @@ class ProductSupplierinfo(models.Model):
     # this is necessary to add a domain to the new product_uom from form view
     product_uom_category_id = fields.Many2one('uom.category', related='product_tmpl_id.uom_po_id.category_id')
     # this is the new editable product uom
-    product_uom = fields.Many2one('uom.uom', ondelete='set null', string='Unit of Measure (Vendor)', store=True, readonly=False, related=False, required=True)
-
-    
-    
-
-
-    
+    product_uom = fields.Many2one('uom.uom', string='Unit of Measure (Vendor)', store=True, readonly=False, related=False, required=True)
