+++ mindesa_purchase/models/user.py
@@ -1,7 +1,7 @@
 from odoo import api, fields, models, _
 
 
-class Partner(models.Model):
-    _inherit = "res.partner"
+class Users(models.Model):
+    _inherit="res.users"
 
     is_rfq_confirm = fields.Boolean(string="Is RFQ confirm")
