+++ mindesa_stock/models/__init__.py
@@ -1,3 +1,1 @@
-# -*- coding: utf-8 -*-
 from . import sale_order
-from . import stock_location
